backend/requirements.txt
# Основные
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
pyyaml==6.0.1
python-dotenv==1.0.0

# Для системной информации
psutil==5.9.6